requirements.txt
asciimatics==1.14.0
asgiref==3.7.2
atomicwrites==1.4.1
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==2.0.12
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.4
defusedxml==0.8.0rc2
Django==3.2
django-bootstrap5==22.2
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
djoser==2.1.0
drf-yasg==1.21.7
Faker==12.0.1
flake8==5.0.4
flake8-broken-line==0.6.0
flake8-complex-f-strings==0.1.0
flake8-isort==6.0.0
flake8-plugin-utils==1.3.2
flake8-return==1.2.0
future==0.18.3
idna==3.4
importlib-metadata==6.7.0
inflection==0.5.1
iniconfig==2.0.0
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
mccabe==0.7.0
mixer==7.2.2
mypy==1.3.0
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
pep8-naming==0.13.3
Pillow==9.3.0
platformdirs==3.7.0
pluggy==0.13.1
py==1.11.0
pycodestyle==2.9.1
pycparser==2.21
pyfiglet==0.8.post1
pyflakes==2.5.0
PyJWT==2.1.0
pyparsing==3.0.9
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
requests==2.26.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.2
sqlparse==0.4.4
toml==0.10.2
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==1.26.17
wcwidth==0.2.6
webcolors==1.11.1
yapf==0.40.1
zipp==3.15.0
